+++ tasks/makejs.js
@@ -1,5 +1,4 @@
-/* eslint  one-var: 0, prefer-arrow-callback: 0, import/no-extraneous-dependencies: 0,
-  semi-style: 0 */
+/* eslint  one-var: 0, import/no-extraneous-dependencies: 0, semi-style: 0 */
 
 'use strict';