+++ tasks/makeprivatepackage.js
@@ -1,16 +1,15 @@
-/* eslint one-var: 0, prefer-arrow-callback: 0, import/no-extraneous-dependencies: 0,
-  semi-style: 0 */
+/* eslint one-var: 0, import/no-extraneous-dependencies: 0, semi-style: 0 */
 
 'use strict';
 
 // -- Node modules
-const { series } = require('gulp')
-    , fs         = require('fs')
+const { src, dest, parallel } = require('gulp')
+    , fs = require('fs')
     ;
 
 
 // -- Local modules
-const config  = require('./config')
+const config = require('./config')
     ;
 
 
@@ -19,6 +18,7 @@ const config  = require('./config')
 
 // -- Local constants
 const { dist } = config
+    , { index } = config
     ;
 
 
@@ -27,27 +27,35 @@ const { dist } = config
 
 // -- Gulp Private Tasks
 
-function makeprivatepackage(cb) {
-  fs.readFile('./package.json', 'utf8', function(error, data) {
+function copypackagejson(done) {
+  fs.readFile('./package.json', 'utf8', (error, data) => {
     if (error) { throw error; }
     const obj = JSON.parse(data);
-    obj.main = {};
+    obj.bin = {};
     obj.scripts = {};
     obj.dependencies = {};
     obj.devDependencies = {};
     obj.private = true;
+    obj.husky = {};
 
     // Write the updated package.json:
     fs.writeFile(`${dist}/package.json`, JSON.stringify(obj, null, 2), 'utf8', (err) => {
       if (err) {
         throw err;
       }
-      cb();
+      done();
     });
   });
+}
+
+// Copies the index.
+function copyindex() {
+  return src(index)
+    .pipe(dest(dist))
+  ;
 }
 
 
 // -- Gulp Public Task(s)
 
-module.exports = series(makeprivatepackage);
+module.exports = parallel(copypackagejson, copyindex);